refactor : template 파일 분리
